Reorg in the server files
