+++ maboss/maboss_server/comm.py
@@ -21,11 +21,8 @@
 # Date: May-December 2018
 
 from __future__ import print_function
-import os, sys, time, signal, socket
-import maboss.maboss_server.atexit
-import maboss.result
-from io import StringIO
-import numpy as np
+
+from .result import Result
 #
 # MaBoSS Communication Layer
 #
@@ -131,7 +128,7 @@ class DataStreamer:
         if hints:
             verbose = "verbose" in hints and hints["verbose"]
 
-        result_data = ResultData(simulation)
+        result_data = Result(simulation)
         magic = RETURN + " " + MABOSS_MAGIC
         magic_len = len(magic)
         if ret_data[0:magic_len] != magic:
@@ -217,197 +214,3 @@ class DataStreamer:
             header += directive + str(o_offset) + "-" + str(offset-1) + "\n"
 
         return (header, offset)
-
-class ClientData:
-
-    def __init__(self, network = None, config = None, command = 'Run'):
-        self._network = network
-        self._config = config
-        self._command = command
-
-    def getNetwork(self):
-        return self._network
-
-    def getConfig(self):
-        return self._config
-
-    def getCommand(self):
-        return self._command
-
-    def setNetwork(self, network):
-        self._network = network
-
-    def setConfig(self, config):
-        self._config = config
-
-class ResultData(maboss.Result):
-
-    def __init__(self, simulation):
-        maboss.Result.__init__(self, simulation)
-        self._status = 0
-        self._errmsg = ""
-        self._stat_dist = None
-        self._prob_traj = None
-        self._traj = None
-        self._FP = None
-        self._runlog = None
-
-    def setStatus(self, status):
-        self._status = status
-
-    def setErrorMessage(self, errmsg):
-        self._errmsg = errmsg
-
-    def setStatDist(self, data_value):
-        self._stat_dist = data_value
-
-    def setProbTraj(self, data_value):
-        self._prob_traj = data_value
-
-    def setTraj(self, data_value):
-        self._traj = data_value
-
-    def setFP(self, data_value):
-        self._FP = data_value
-
-    def setRunLog(self, data_value):
-        self._runlog = data_value
-
-    def getStatus(self):
-        return self._status
-
-    def getErrorMessage(self):
-        return self._errmsg
-
-    def getStatDist(self):
-        return self._stat_dist
-
-    def getProbTraj(self):
-        return self._prob_traj
-
-    def getTraj(self):
-        return self._traj
-
-    def getFP(self):
-        return self._FP
-
-    def getRunLog(self):
-        return self._runlog
-
-    def get_fp_file(self):
-        return StringIO(self.getFP())
-
-    def get_probtraj_file(self):
-        return StringIO(self.getProbTraj())
-
-    def get_probtraj_dtypes(self):
-
-        first_line = self.getProbTraj().split("\n", 1)[0]
-        cols = first_line.split("\t")
-        nb_states = int((len(cols) - 5)/3)
-        dtype = {"Time": np.float64, "TH": np.float64, "ErrorTH": np.float64, "H": np.float64, "HD=0": np.float64, "State": np.str, "Proba": np.float64, "ErrorProba": np.float64}
-        for i in range(1, nb_states):
-            dtype.update({"State.%d" % i: np.str, "Proba.%d" % i: np.float64, "ErrorProba.%d" % i: np.float64})
-        return dtype
-
-
-class MaBoSSClient:
-    
-    SERVER_NUM = 1 # for now
-
-    def __init__(self, host = None, port = None, maboss_server = None):
-        if not maboss_server:
-            maboss_server = os.getenv("MABOSS_SERVER")
-            if not maboss_server:
-                maboss_server = "MaBoSS-server"
-
-        self._maboss_server = maboss_server
-        self._host = host
-        self._port = port
-        self._pid = None
-        self._mb = bytearray()
-        self._mb.append(0)
-        self._pidfile = None
-
-        if host == None:
-            if port == None:
-                port = '/tmp/MaBoSS_pipe_' + str(os.getpid()) + "_" + str(MaBoSSClient.SERVER_NUM)
-
-            self._pidfile = '/tmp/MaBoSS_pidfile_' + str(os.getpid()) + "_" + str(MaBoSSClient.SERVER_NUM)
-            MaBoSSClient.SERVER_NUM += 1
-
-            try:
-                pid = os.fork()
-            except OSError as e:
-                print("error fork:", e, file=sys.stderr)
-                return
-
-            if pid == 0:
-                try:
-                    args = [self._maboss_server, "--host", "localhost", "-q", "--port", port, "--pidfile", self._pidfile]
-                    os.execvp(self._maboss_server, args)
-                except Exception as e:
-                    print("error while launching '" + self._maboss_server + "'", e, file=sys.stderr)
-                    sys.exit(1)
-
-            self._pid = pid
-            maboss.maboss_server.atexit.register(self.close)
-            server_started = False
-            MAX_TRIES = 20
-            TIME_INTERVAL = 0.1
-            for try_cnt in range(MAX_TRIES):
-                if os.path.exists(self._pidfile):
-                    server_started = True
-                    break
-                time.sleep(TIME_INTERVAL)
-
-            if not server_started:
-                self._pidfile = None
-                raise Exception("MaBoSS server '" + self._maboss_server + "' on port " + port + " not started after " + str(MAX_TRIES*TIME_INTERVAL) + " seconds")
-
-            self._socket = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-            self._socket.connect(port)
-        else:
-            self._socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            self._socket.connect((host, port))
-            
-    def run(self, simulation, hints={}):
-
-        if "check" in hints and hints["check"]:
-            command = CHECK_COMMAND
-        else:
-            command = RUN_COMMAND
-
-        client_data = ClientData(str(simulation.network), simulation.str_cfg(), command)
-
-        data = DataStreamer.buildStreamData(client_data, hints)
-        data = self.send(data)
-
-        return DataStreamer.parseStreamData(data, simulation, hints )
-
-        # return maboss.maboss_server.result.Result(self, simulation, hints).getResultData()
-
-    def send(self, data):
-        self._socket.sendall(data.encode())
-        self._term()
-        SIZE = 4096
-        ret_data = ""
-        while True:
-            databuf = self._socket.recv(SIZE)
-            if not databuf or len(databuf) <= 0:
-                break
-            ret_data += databuf.decode()
-
-        return ret_data
-
-    def _term(self):
-        self._socket.send(self._mb)
-
-    def close(self):
-        if self._pid != None:
-            #print("kill", self._pid)
-            os.kill(self._pid, signal.SIGTERM)
-            if self._pidfile:
-                os.remove(self._pidfile)
-            self._pid = None
-        self._socket.close()
